update filenames for Intel arch files
Move away from including microarchitecture name in Intel filenames to just
including model/family in the filenames. For example, move from Broadwell_4F to
Intel_06_4F
