+++ src/variorum/Intel/CMakeLists.txt
@@ -9,14 +9,14 @@ set(variorum_intel_headers
   ${CMAKE_CURRENT_SOURCE_DIR}/power_features.h
   ${CMAKE_CURRENT_SOURCE_DIR}/thermal_features.h
   ${CMAKE_CURRENT_SOURCE_DIR}/misc_features.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/SandyBridge_2A.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/SandyBridge_2D.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/IvyBridge_3E.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/Haswell_3F.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/Broadwell_4F.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/Skylake_55.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/KabyLake_9E.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/IceLake_6A.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_2A.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_2D.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_3E.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_3F.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_4F.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_55.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_9E.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_6A.h
   ${CMAKE_CURRENT_SOURCE_DIR}/variorum_cpuid.h
   CACHE INTERNAL "")
 
@@ -26,14 +26,14 @@ set(variorum_intel_sources
   ${CMAKE_CURRENT_SOURCE_DIR}/power_features.c
   ${CMAKE_CURRENT_SOURCE_DIR}/thermal_features.c
   ${CMAKE_CURRENT_SOURCE_DIR}/misc_features.c
-  ${CMAKE_CURRENT_SOURCE_DIR}/SandyBridge_2A.c
-  ${CMAKE_CURRENT_SOURCE_DIR}/SandyBridge_2D.c
-  ${CMAKE_CURRENT_SOURCE_DIR}/IvyBridge_3E.c
-  ${CMAKE_CURRENT_SOURCE_DIR}/Haswell_3F.c
-  ${CMAKE_CURRENT_SOURCE_DIR}/Broadwell_4F.c
-  ${CMAKE_CURRENT_SOURCE_DIR}/Skylake_55.c
-  ${CMAKE_CURRENT_SOURCE_DIR}/KabyLake_9E.c
-  ${CMAKE_CURRENT_SOURCE_DIR}/IceLake_6A.c
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_2A.c
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_2D.c
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_3E.c
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_3F.c
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_4F.c
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_55.c
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_9E.c
+  ${CMAKE_CURRENT_SOURCE_DIR}/Intel_06_6A.c
   ${CMAKE_CURRENT_SOURCE_DIR}/variorum_cpuid.c
   CACHE INTERNAL "")
